+++ EssentialFeed/EssentialFeed/Feed Cache/Infrastructure/CoreData/CoreDataFeedStore.swift
@@ -65,48 +65,3 @@ public final class CoreDataFeedStore: FeedStore {
 	}
 
 }
-
-@objc(ManagedCache)
-private class ManagedCache: NSManagedObject {
-	@NSManaged var timestamp: Date
-	@NSManaged var feed: NSOrderedSet
-	
-	static func find(in context: NSManagedObjectContext) throws -> ManagedCache? {
-		let request = NSFetchRequest<ManagedCache>(entityName: entity().name!)
-		request.returnsObjectsAsFaults = false
-		return try context.fetch(request).first
-	}
-	
-	static func newUniqueInstance(in context: NSManagedObjectContext) throws -> ManagedCache {
-		try find(in: context).map(context.delete)
-		return ManagedCache(context: context)
-	}
-	
-	var localFeed: [LocalFeedImage] {
-		return feed.compactMap { ($0 as? ManagedFeedImage)?.local }
-	}
-}
-
-@objc(ManagedFeedImage)
-private class ManagedFeedImage: NSManagedObject {
-	@NSManaged var id: UUID
-	@NSManaged var imageDescription: String?
-	@NSManaged var location: String?
-	@NSManaged var url: URL
-	@NSManaged var cache: ManagedCache
-	
-	static func images(from localFeed: [LocalFeedImage], in context: NSManagedObjectContext) -> NSOrderedSet {
-		return NSOrderedSet(array: localFeed.map { local in
-			let managed = ManagedFeedImage(context: context)
-			managed.id = local.id
-			managed.imageDescription = local.description
-			managed.location = local.location
-			managed.url = local.url
-			return managed
-		})
-	}
-	
-	var local: LocalFeedImage {
-		return LocalFeedImage(id: id, description: imageDescription, location: location, url: url)
-	}
-}
